pyproject.toml
[tool.poetry]
name = "pyOPENDART"
version = "0.2.0-alpha.3"
description = "전자공시시스템 DART 파이썬 API (for Humans)"
authors = ["Seonghyeon Kim <self@seonghyeon.dev>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/NovemberOscar/pyOPENDART"
keywords = ["data", "market", "stock", "korea", "investment", "kospi", "kosdaq"]
classifiers=[
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Korean",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Office/Business :: Financial :: Investment",
]

[tool.poetry.dependencies]
python = "^3.7.1"
requests = "^2.25.1"
furl = "^2.1.0"
python-dateutil = "^2.8.1"
pandas = "^1.2.1"
xmltodict = "^0.12.0"
httpx = {version = "^0.16.1", optional = true}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
notebook = "^6.2.0"
mkdocs = "^1.1.2"
mkdocs-material = "^6.2.8"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"

[tool.poetry.extras]
httpx = ["httpx"]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs         # exclude a few common directories in the
  | \.git          # root of the project
  | \.hg
  | \.mypy_cache
  | \.tox
  | buck-out
  | build
  | dist
  | tests
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
